refactor(ibverbs): wrap DeviceContext, ProtectionDomain, MemoryRegion in Arc
Currently we are using lifetime mark for them, but this would introduce
difficulty for user to encap them in a single struct (which would be a
self-referential struct). Both of them would only be used in control
path, so change to wrap them in Arc wouldn't be a big performance issue,
but could benefit users a lot.

Signed-off-by: Luke Yue <lukedyue@gmail.com>